refactor(auth): update import paths for validation schemas and remove unused data file
