+++ web-app/src/app/(landing)/(auth)/signin/data.ts
@@ -1,8 +1,7 @@
 import { z } from "zod";
 
+import { emailSchema, passwordSchema } from "@/app/(auth)/data";
 import { FormData } from "@/lib/form";
-
-import { emailSchema, passwordSchema } from "../data";
 
 const signInFormSchema = z.object({
   email: emailSchema,
